refactor: divide the setup guide page
