+++ packages/mask/src/components/DataSource/useNextID.ts
@@ -1,15 +1,15 @@
 import { useAsyncRetry } from 'react-use'
-import type { PersonaIdentifier } from '@masknet/shared-base'
-import { queryExistedBindingByPersona, NextIDPlatform, queryExistedBindingByPlatform } from '@masknet/web3-providers'
+import type { NextIDPlatform, PersonaIdentifier } from '@masknet/shared-base'
+import Services from '../../extension/service'
 
 export const usePersonaBoundPlatform = (personaIdentifier: PersonaIdentifier) => {
     useAsyncRetry(() => {
-        return queryExistedBindingByPersona(personaIdentifier)
+        return Services.NextID.queryExistedBindingByPersona(personaIdentifier)
     }, [personaIdentifier.toText()])
 }
 
 export const useNextIDBoundByPlatform = (platform: NextIDPlatform, identity: string) => {
     useAsyncRetry(() => {
-        return queryExistedBindingByPlatform(platform, identity)
+        return Services.NextID.queryExistedBindingByPlatform(platform, identity)
     }, [platform, identity])
 }
