+++ packages/mask/src/components/DataSource/usePersonaConnectStatus.ts
@@ -4,7 +4,7 @@ import { useMemo } from 'react'
 import Services from '../../extension/service'
 import { currentSetupGuideStatus } from '../../settings/settings'
 import { activatedSocialNetworkUI } from '../../social-network'
-import { SetupGuideStep } from '../InjectedComponents/SetupGuide'
+import { SetupGuideStep } from '../InjectedComponents/SetupGuide/types'
 import { useLastRecognizedIdentity } from './useActivatedUI'
 import { useMyPersonas } from './useMyPersonas'
