+++ packages/mask/src/components/InjectedComponents/SetupGuide/types.ts
@@ -1,0 +1,5 @@
+export enum SetupGuideStep {
+    FindUsername = 'find-username',
+    PinExtension = 'pin-extension',
+    Close = 'close',
+}
