+++ packages/mask/src/extension/background-script/SocialNetworkService.ts
@@ -4,7 +4,8 @@ import { requestSNSAdaptorPermission } from '../../social-network/utils/permissi
 
 import { currentSetupGuideStatus } from '../../settings/settings'
 import stringify from 'json-stable-stringify'
-import { SetupGuideStep } from '../../components/InjectedComponents/SetupGuide'
+import { PersonaIdentifier } from '@masknet/shared-base'
+import { SetupGuideStep } from '../../components/InjectedComponents/SetupGuide/types'
 import type { PersonaIdentifier } from '@masknet/shared-base'
 import { delay } from '@dimensiondev/kit'
