+++ packages/mask/src/social-network/types.ts
@@ -309,6 +309,8 @@ export namespace SocialNetworkUI {
         }
         export interface NextIDConfig {
             enable?: boolean
+            platform: string
+            collectVerifyPost: (keyword: string) => string | null
         }
     }
 }
