+++ packages/shared-base/src/index.ts
@@ -15,3 +15,4 @@ export { DashboardRoutes } from './Routes/DashboardRoutes'
 export * from './Messages/Mask'
 export * from './Results'
 export * from './convert'
+export * from './NextID/type'
