+++ packages/web3-providers/package.json
@@ -8,7 +8,6 @@
     "@masknet/web3-constants": "workspace:*",
     "@masknet/web3-shared-base": "workspace:*",
     "@masknet/web3-shared-evm": "workspace:*",
-    "@masknet/shared-base": "workspace:*",
     "bignumber.js": "9.0.2",
     "date-fns": "^2.28.0",
     "json-stable-stringify": "^1.0.1",
