+++ packages/web3-providers/src/index.ts
@@ -11,7 +11,6 @@ import { TokenPriceAPI } from './token-price'
 
 export * from './types'
 export * from './opensea/utils'
-export * from './NextID'
 
 export const OpenSea = new OpenSeaAPI()
 export const Rarible = new RaribleAPI()
